pyproject.toml
[project]
name = "mood-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
